requirements.txt
Django
markdown
django-filter
djangorestframework